Kategori crud işlemleri gerçekleştirildi.

+++ BlogFall/Areas/Admin/Controllers/PostsController.cs
@@ -1,17 +1,22 @@
 ﻿using BlogFall.Areas.Admin.ViewModel;
+using BlogFall.Attributes;
 using BlogFall.Models;
 using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Net;
 using System.Web;
 using System.Web.Mvc;
 
 namespace BlogFall.Areas.Admin.Controllers
 {
+    [Breadcrumb("Yazılar")]
     public class PostsController : AdminBaseController
     {
         // GET: Admin/Posts
+        [Breadcrumb("İndeks")]
         public ActionResult Index()
         {
             return View(db
@@ -24,12 +29,19 @@ namespace BlogFall.Areas.Admin.Controllers
         public ActionResult Delete(int id)
         {
             var post = db.Posts.Find(id);
+
+            if (post == null)
+            {
+                return HttpNotFound();
+            }
+
             db.Posts.Remove(post);
             db.SaveChanges();
 
             return Json(new { success=true});
         }
 
+        [Breadcrumb("Düzenle")]
         public ActionResult Edit(int id)
         {
             ViewBag.CategoryId = new SelectList(db.Categories.ToList(), "Id", "CategoryName");
@@ -48,6 +60,7 @@ namespace BlogFall.Areas.Admin.Controllers
         [HttpPost]
         [ValidateInput(false)]
         [ValidateAntiForgeryToken]
+        [Breadcrumb("Düzenle")]
         public ActionResult Edit(PostEditViewModel model)
         {
             if (ModelState.IsValid)
@@ -65,6 +78,7 @@ namespace BlogFall.Areas.Admin.Controllers
             return View();
         }
 
+        [Breadcrumb("Yeni")]
         public ActionResult New()
         {
             ViewBag.CategoryId = new SelectList(db.Categories.ToList(), "Id", "CategoryName");
@@ -75,6 +89,7 @@ namespace BlogFall.Areas.Admin.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Breadcrumb("Yeni")]
         [ValidateInput(false)]//html taglerini kontrol güvenliğini iptal etme
         public ActionResult New(PostEditViewModel model)
         {
@@ -99,6 +114,24 @@ namespace BlogFall.Areas.Admin.Controllers
             ViewBag.CategoryId = new SelectList(db.Categories.ToList(), "Id", "CategoryName");
 
             return View("Edit", new PostEditViewModel());
+        }
+
+        [HttpPost]
+        public ActionResult AjaxImageUpload(HttpPostedFileBase file)
+        {
+            if (file == null || file.ContentLength == 0 || !file.ContentType.StartsWith("image/"))//yüklenen dosya resim olmalı,
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
+            //diğer durumlarda dosya yüklenir
+            var saveFolderPath = Server.MapPath("~/Upload/Posts");//harddisk te fiziksel olarak klasörün bulunduğu yolu döndürür.
+            var ext = Path.GetExtension(file.FileName);//uzantısı ne onu alır
+            var saveFileName = Guid.NewGuid().ToString() + ext;
+            var saveFilePath = Path.Combine(saveFolderPath, saveFileName);
+            file.SaveAs(saveFilePath);//kaydediyor. buse.jpg gibi bir yol elde ediyoruz bu adımlarla
+
+            return Json(new { url = Url.Content("~/Upload/Posts/" + saveFileName) });
         }
     }
 }
